+++ src/frontend/components/SøknadsSteg/Oppsummering/OppsummeringSteg/OmBarnet/OmBarnetOppsummering.tsx
@@ -14,14 +14,17 @@ import {
     IBarnMedISøknad,
 } from '../../../../../typer/person';
 import { formaterDato } from '../../../../../utils/dato';
-import { barnetsNavnValue, landkodeTilSpråk } from '../../../../../utils/visning';
+import {
+    barnetsNavnValue,
+    formaterDatoMedUkjent,
+    landkodeTilSpråk,
+} from '../../../../../utils/visning';
 import SpråkTekst from '../../../../Felleskomponenter/SpråkTekst/SpråkTekst';
 import { OmBarnetSpørsmålsId, omBarnetSpørsmålSpråkId } from '../../../OmBarnet/spørsmål';
 import { useOmBarnet } from '../../../OmBarnet/useOmBarnet';
 import { StyledOppsummeringsFeltGruppe } from '../../Oppsummering';
 import { OppsummeringFelt } from '../../OppsummeringFelt';
 import Oppsummeringsbolk from '../../Oppsummeringsbolk';
-import { formaterDatoMedUkjent } from '../../utils';
 import AndreForelderOppsummering from './AndreForelderOppsummering';
 
 interface Props {
