--- src/frontend/components/SøknadsSteg/Oppsummering/utils.ts
@@ -1,8 +1,0 @@
-import { AlternativtSvarForInput, DatoMedUkjent } from '../../../typer/person';
-import { formaterDato } from '../../../utils/dato';
-
-export const formaterDatoMedUkjent = (datoMedUkjent: DatoMedUkjent, tekstForUkjent) => {
-    return datoMedUkjent === AlternativtSvarForInput.UKJENT
-        ? tekstForUkjent
-        : formaterDato(datoMedUkjent);
-};
