+++ src/frontend/components/SøknadsSteg/Utvidet-DinLivssituasjon/useTidligereSamboer.ts
@@ -1,11 +1,11 @@
 import { ESvar, ISODateString } from '@navikt/familie-form-elements';
 import { ISkjema, useFelt, useSkjema } from '@navikt/familie-skjema';
 
+import useDatovelgerFelt from '../../../hooks/useDatovelgerFelt';
+import useDatovelgerFeltMedUkjent from '../../../hooks/useDatovelgerFeltMedUkjent';
 import useInputFelt from '../../../hooks/useInputFelt';
 import useInputFeltMedUkjent from '../../../hooks/useInputFeltMedUkjent';
 import { DatoMedUkjent } from '../../../typer/person';
-import useDatovelgerFelt from '../OmBarnet/useDatovelgerFelt';
-import useDatovelgerFeltMedUkjent from '../OmBarnet/useDatovelgerFeltMedUkjent';
 import { TidligereSamboerSpørsmålId } from './spørsmål';
 
 export interface ITidligereSamboerFeltTyper {
