+++ src/frontend/hooks/useLanddropdownFelt.tsx
@@ -4,8 +4,8 @@ import { Alpha3Code } from 'i18n-iso-countries';
 
 import { feil, FeltState, ok, useFelt } from '@navikt/familie-skjema';
 
-import { ISøknadSpørsmål } from '../../../typer/søknad';
-import SpråkTekst from '../../Felleskomponenter/SpråkTekst/SpråkTekst';
+import SpråkTekst from '../components/Felleskomponenter/SpråkTekst/SpråkTekst';
+import { ISøknadSpørsmål } from '../typer/søknad';
 
 const useLanddropdownFelt = (
     søknadsfelt: ISøknadSpørsmål<Alpha3Code | ''>,
