+++ src/frontend/utils/språk.ts
@@ -1,22 +1,4 @@
-export enum Årsak {
-    SEPARERT = 'SEPARERT',
-    SKILT = 'SKILT',
-    BRUDD_SAMBOER = 'BRUDD_SAMBOER',
-    BODD_ALENE = 'BODD_ALENE',
-    ENKE_ENKEMANN = 'ENKE_ENKEMANN',
-    FENGSEL_VARETEKT = 'FENGSEL_VARETEKT',
-    BRUDD_GIFT = 'BRUDD_GIFT',
-}
-
-export const muligeÅrsaker: Årsak[] = [
-    Årsak.SEPARERT,
-    Årsak.SKILT,
-    Årsak.BRUDD_SAMBOER,
-    Årsak.BODD_ALENE,
-    Årsak.ENKE_ENKEMANN,
-    Årsak.FENGSEL_VARETEKT,
-    Årsak.BRUDD_GIFT,
-];
+import { Årsak } from '../typer/søknad';
 
 export const toÅrsakSpråkId = (årsak: Årsak): string => {
     switch (årsak) {
